feat!: add disambiguation for Valkey/Redis in storage layer (#77)
BREAKING CHANGE: This modifies the ID used to identify clusters
in the Autoscaler state storage layer. When this version is
first deployed to an existing environment that is managed by
the Autoscaler, it may trigger a single additional scaling
operation during a cooldown period or while a previously
started scaling operation is ongoing.